requirements.txt
streamlit
mistralai
snowflake-connector-python
python-dotenv
requests
pandas
numpy